requirements-dev.txt
dj-database-url==0.5.0
dj-static==0.0.6
Django==1.11.11
djangorestframework==3.7.7
nltk==3.2.5
python-decouple==3.1
pytz==2018.3
six==1.11.0
static3==0.7.0
